Add "'cobalt_code': 1" to dom.gyp, fix Windows build
To make cobalt_code 1 work, the following changes are made:
- Added MSVC_PUSH_DISABLE_WARNING 4355 to anywhere "this" is used in base
  member initializer list.
- Disabled warning C4702: unreachable code in xtree.
- Added static_cast to all places where implicit conversion will cause warnings,
  or change variable type if it makes sense.
- Fixed all unreferenced parameter warnings.

b/22613037

+++ cobalt/dom/node_list.cc
@@ -34,7 +34,7 @@ scoped_refptr<NodeList> NodeList::CreateWithChildren(
 
 unsigned int NodeList::length() {
   MaybeRefreshCollection();
-  return cached_collection_.size();
+  return static_cast<unsigned int>(cached_collection_.size());
 }
 
 scoped_refptr<Node> NodeList::Item(unsigned int item) {
